Update fix for 102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	unsigned long long curr, prev, sum;
+	long int curr, prev, sum;
 	int i;
 
 	printf("1, 2, ");
@@ -17,7 +17,7 @@ int main(void)
 	for (i = 2; i < 50; i++)
 	{
 		sum = curr + prev;
-		printf("%llu", sum);
+		printf("%ld", sum);
 		prev = curr;
 		curr = sum;
